requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
python-multipart>=0.0.6
requests>=2.28.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
ollama>=0.1.7
# Simplified agent framework - we'll implement custom agents
openai>=1.0.0
langchain-core>=0.1.0
langchain-ollama>=0.1.0
